Allow to show source through browser
When someone bumps into the camera setup and you want to re-calibrate,
you always needed to find a terminal to start the show-source
script, which was tedious.

We turn the show-source script into a cgi script, and hook it up with
the letter 's'. So pressing 's' in the browser will fire up the camera
source window and allows to quickly re-adjust the camera setup.