Allow to reindex through browser
When remotely copying dynamic content to the server, one needs to run
`reindex` to merge old and new available files. This up to now
required finding a console to run the command. Since reindexing is
harmles, and it's convenient to be able to trigger a reindexing run
through the browser, we hook up reindexing with pressing 'r' in the
browser.