Bring cgi failure/success to screen and console.log
Without any visual feedback, it was hald to tell if a cgi call failed
or succeeded, or if the key got registered at all.

We add the cgi failure/success to the screen to make it visible right
away, and also to the console.log, so one can look up failures after
the fact.

Closes: #5