Have a more aggressive default destroyOffset
We assume that actors have their origin on the main image. This allows
to trim the destroyOffset considerably, which leads to way earlier
destroying and freeing up of CPU resources.

If the actor does not have its origin on the main image, it's expected
to manually set the destroyOffset accordingly.