Scan for dedicated rect instead of paper contour
The paper contour is not too reliable, as it depends on the external
factors. One such factor is the background, which might have features
that let the paper sheet appear to have outward finger-like
extensions. Another factor is the printer paper feeding
mechanism. Some paper feeding mechanisms allow for small random
rotations, which make it hard to reliably correlate the actor drawing
within the scanned paper contour. These and further factors are in the
way of a reliable and spot-on detection.

To improve detection quality, we add a 2mm rectangle to the border of
actor images and detect this (instead of the paper edges). As this
rectangle is always surrounded by the white paper, so this method is
no longer as dependent of the background of the scanning plane. And
this method also fixes the relative position between the detected
contour and the relevant image content.

With this change we arrive at reliable, spot-on detection even on old
640x480 cameras.