Stop locale unsetting when showing images in debug mode
`cv2.imshow` seems to set `LC_NUMERIC` to empty, and hence unsets the
locale when in debug mode. With the unset locale, calls to `inkscape`
fail as point characters get used as decimal points when passing
floats to inkscape (instead of localized decimal points).

So we reset the locale after calls to `cv2.imshow` to work around the
unsetting.