Improve error message for unknown QR codes
When scanning a QR code for which there is no actor, the error message
complained about missing the actor mask. While this was technically
correct, is was misleading, as the root cause is that either the scene
does not exist (and hence the command is invalid) or that the scene
does not have the requested actor. We now give more precise error
messages in these two cases.