Allow to directly configure used OpenCV camera specification
This helps in situations when a plain `/dev/video0` is not good enough
to open a camera for OpenCV (E.g.: When using a GStreamer source).

We require the new `image:` prefix for image sources. This makes the
more relevant case (raw OpenCV camera specification) the default and
pushes the special case (image source) behind a prefix.