Allow different qr-pixels for commands and actors
As command pages see heavier use, they are typically laminated.
Laminated pages however are typically very reflective and show
reflected light points that would be invisible on plain paper, which is
in the way of reliable QR code detection.

To get reliable QR code detection even for laminated pages, we add a
dedicated qr-pixel for commands, which allows to move the QR code
closer to the center (where the chances of reflections are smaller)
for command pages.