frontend: Default to browser UI language
If no language was selected, we defaulted to English. This is not a
good choice for countries where English is not the primary language.

The best choice would probably be using the `Accept-Language` header.
But this is not available in plain JavaScript. So it would require
another CGI to shoehorn it back to the javascript part. Picking the
language of the browser UI however is a good-enough default for now
and requires no additional tooling, so we choose that as default for
now.