localization: Add empty English translation
Since development happens in English, there is no need to translate
any string for English UI. But the frontend on purpose requests the
localization data regardless. This request received a 404 for English
for now. As that looks like something is wronge, we add an empty
translation file to make the request a 200.