services/continuous-scanning: Reset timer after bailout
By not resetting the bailout timer, the bailout gave ~1 second to bring
the camera pipeline back into shape. If that was not enough, the next
bailout got triggered and so on. This did work for pipeline where
resetting is quick, but failed for pipelines, where the resetting takes
longer.

So we now reset the bailout timer after a bailout. This should give
enough time to reset the camera pipeline.