highway: Limit tire source to scanned image
Due to scaling, x2 and y2 need not be on the scanned image, but may lie
outside. In such cases the last row/column got repeated when blitting
to the target, which gave duplicated lines/rows, which made tires look
un-round. So we limit the source to the scanned image.