ballons: Reduce to a single sample actors
We drop the red, and blue sample balloons and only keep the yellow one.
A single sample actor is enough for a scene. The sample actors should
not give a pretty scene and shift focus to them, but they should only
bring some life to the screen until the users scanned their own
drawings.