Base64 encode binary parameters
argparse is trying to see if parameters start with a dash. This helps
argparse to decide how to process the parameters. During these checks,
argparse assumes that each argument is a string, as it's doing
`<string> in <parameter>`. This obviously fails, if parameter is a
binary value. So we automatically encode binary parameters harvested
from cgi.

This allows to keep using argparse, while still being able to call the
scripts from the command-line and still being able to provide `curl`
friendly cgi-bins. As this en-/decoding is only happening behind the
scenes, we can get rid of it at some point in the future without much
disruption, if that has been fixed in argparse.