frontend: Add 1px transparend border to message icons
A ghosting effect was visble when the icons were rendered on a screen
that required scaling. For the `ok` icon, pixel from the right most
columns were vilible on the left most. Similarly, for the pause icon
and top/bottom rows. By adding a 1 pixel transparent border, only
transparent pixels get carried over to the opposite edge, which makes
the ghosting effect dissappear.