Indexer: Add key for each actor in dynamic json files
We indexed each scan of an actor. This means that images stay in the
index, even if the static artifacts (Javascript file, mask, ...) are
gone. And also the json file could not get used to determine all
available actors. We switch to indexing all dynamic content for which
we have static assets. This solves the above two issues.