Revert "Indexer: Add key for each actor in dynamic json files"
Using the actors.json file to determine which PDFs to advertize, proved
to be the wrong track. First, it imposes extra work for each reindex,
which is unwarranted. And then it does not contain information about
the actor variants. And it cannot contain this information without
parsing the SVGs. So we added a dedicated `actors-varians.json` in
80afb03 which better fits this purpose. So we revent the now no longer
needed generation of keys with empty content.

This reverts commit 2381033716782ac80b95f931b2ee61a1f8cbc31e.