Move `scenes.json` from `dynamic` to `scenes` directory
The file is not a perfect fit for the `scenes` directory. But as the
file does not depend on any runtime aspects or runtime files and as
it's not meant to be messed with by users or the site, haveing it under
`dynamic` is even worse. So we move it into the `scenes` directory.