input: Make ` ` pause, move scan to `s`, and show-source to `i`
The space key is commonly used as pause. So we follow that lead. The
scan action (which was mapped to space previously) moves to `s`, which
is a natural fit. Therefor, show-source (previously mapped to `s`)
needs to move away. Since the source is also an input, we move it
to `i`.