frontend: greeter: Drop icon for screen-tap message
Previously, we had two `i` icons. One for the welcome message and one
for the screen-tap message. This looked crowded. Also, the screen-tap
message was up only 5 seconds, which was rather short for reading the
whole message.

By switching the the screen-tap message to `none`, the icon disappears,
and the message stays up longer (10 seconds).