frontend: actor-manager: Don't pick first encountered untried actor
By picking the first encountered untried actor, the sequence of actors
was more predictable than we'd want for the demo site.

After picking the config, we build 3 sets. The set of untried variants
that are the first variant in their actors's list, the set of untried
variants, and all variants. The second is only used for the `latest`
config variant. Then the first non-empty of the lists is picked, and a
random variant is chosen.

This makes the actor choice more random, while still prefering fresh
ones.