frontend: settings: Prepare Settings dialog when buttons become visible
On the one hand, this starts downloading the dynamic content (scene
previews, language data, ...) early when the tab is not busy anyways
and thereby helps to show the user a fully populated dialog by the time
the user clicks on the "Settings" button.

On the other hand, by initializing the "Help" tab already before the
Setting get opened, we can keep track of the last failed image in the
Help tab even if the user had not yet opened the Settings page for the
first time when the upload failed.