scan: Drop scanarium parameter from re-rectifying
The signature of `rectify_to_biggest_rect` does not need an extra
Scanarium instance passed, and hence the call failed and lead to an
undefined error instead of a proper error about failing to find the
rectangle.

We drop the parameter, and now the error message is correct again
and better guides users.