frontend: url-parameters: When following new settings, reload page
When setting an URL parameter with `follow` being true, the intention
is to reload the page even if the URL did not change.

However, we instead passed `false` as the first parameter, which means
to /not/ reload the URL if there was no change. The URL got reloaded
nonetheless due to an unrelated bug (`updateLocation` compared the old
query string to the new full url, which are conceptionally different).

As we're about to fix `updateLocation`'s but we need to adjust the
passed first parameter to properly express the intention.