frontend: message-manager: Keep track of last evicted message uuid
The uuid of the most recently message evicted from screen will help
in follow-up commits to decide which messages to show after an
automatic reload.

This is not a fully perfect system, as it may happen that an older
message (e.g.: from a failure) is still on the screen when a newer
one gets evicted. But it will allow to kill the main pain points.