frontend: command-processor: Add `replay` mode
In replay mode, the commands are processed as usual, and messages get
added, but no actions (changing urls, messing with url parameters, ...)
are taken. Only recently scanned actors are added regardless.

This is useful when reloading the window during play, and allows to
replay seen commands to regain (parts of) the on-screen messages.