frontend: pause-manager: Run code to pause only when not yet paused
The `pause` method ran the code to puse the game regardless of whether
it was already paused or not. So when adding the call to `pause` upon
opening the Settings window, the pause code ran twice (Once for tapping
the screen to show the `Settings` button and once for pressing the
`Settings` button) and the `Pause` message appeared twice on the
screen.

By running the pause code only if the game is not paused, we remove
this duplicated message.