frontend: If the WebGL context got lost, skip adding message buttons
On tablets with scarce memory (Amazon Fire Tablet 7), the webgl context
routinely gets lost, when uploading a freshly taken image.
If additionally the uploading fails, a help button was added to the
message. As the message still existed, we received get a fresh button
on a dark screen, which looked wrong.
So we guard the adding of buttons by a check for the webgl context.