Allow to hide actors
If an actor directory contains a file with name `hidden`, the actor is
hidden. Hidden actors are not advertised in `actor-variants.json`
(hence are not visible from Scanarium's PDF tab in the settings
window), and they are not included in the collections of PDF files
(E.g.: The pdf containing all actors)

Hidden actors can be used in Scanarium like any other actor. They scan
properly and act in the scene just as all the other actors do.

Hiding actors allows to have actors that might be useful for a niche
group of users, while still not advertising them to the general public.

The `hidden` file is a plaintext fil ad should contain the reason, why
this actor is hidden.